+++ CHANGELOG.md
@@ -5,8 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-
 ## [UNRELEASED]
+
+### Fixed
+- Add datetime: null when an item has start_datetime and end_datetime but no datetime. Fixes (#158)
+
 
 ##[v0.9.9]